Updated examples and clean up pages
